rearranged files into different folders for organization
