+++ RazorForms/Generators/Elements/IInputBlockWrapperGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using RazorForms.Options;
 
-namespace RazorForms.Generators;
+namespace RazorForms.Generators.Elements;
 
 public interface IInputBlockWrapperGenerator : IOutputGenerator<IFormComponentOptions>
 {
