+++ RazorForms/Generators/Elements/InputBlockWrapperGenerator.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Razor.TagHelpers;
 using RazorForms.Options;
 using RazorForms.TagHelpers;
 
-namespace RazorForms.Generators;
+namespace RazorForms.Generators.Elements;
 
 public class InputBlockWrapperGenerator : OutputGeneratorBase<IFormComponentOptions>, IInputBlockWrapperGenerator
 {
