+++ RazorForms/Generators/Elements/LabelGenerator.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Razor.TagHelpers;
 using RazorForms.Options;
 using RazorForms.TagHelpers;
 
-namespace RazorForms.Generators;
+namespace RazorForms.Generators.Elements;
 
 public class LabelGenerator : OutputGeneratorBase<IFormComponentOptions>, ILabelGenerator
 {
