+++ RazorForms/Generators/Inputs/IInputGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using RazorForms.Options;
 
-namespace RazorForms.Generators;
+namespace RazorForms.Generators.Inputs;
 
 public interface IInputGenerator : IOutputGenerator<IFormComponentOptions>
 {
