+++ RazorForms/Generators/Inputs/ISelectGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using RazorForms.Options;
 
-namespace RazorForms.Generators;
+namespace RazorForms.Generators.Inputs;
 
 public interface ISelectGenerator : IOutputGenerator<IFormComponentOptions>
 {
