+++ RazorForms/TagHelpers/RazorFormsTagHelperBase.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.TagHelpers;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using RazorForms.Generators;
+using RazorForms.Generators.Elements;
 using RazorForms.Options;
 
 namespace RazorForms.TagHelpers;
