+++ RazorForms/TagHelpers/SelectInputTagHelper.cs
@@ -5,6 +5,8 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using Microsoft.Extensions.Logging;
 using RazorForms.Generators;
+using RazorForms.Generators.Elements;
+using RazorForms.Generators.Inputs;
 using RazorForms.Options;
 
 namespace RazorForms.TagHelpers;
