+++ RazorForms/TagHelpers/TextAreaInputTagHelper.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using RazorForms.Generators;
+using RazorForms.Generators.Elements;
+using RazorForms.Generators.Inputs;
 using RazorForms.Options;
 
 namespace RazorForms.TagHelpers;
