+++ RazorForms/TagHelpers/TextInputTagHelper.cs
@@ -12,6 +12,8 @@ using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.AspNetCore.Razor.TagHelpers;
 using Microsoft.Extensions.Logging;
 using RazorForms.Generators;
+using RazorForms.Generators.Elements;
+using RazorForms.Generators.Inputs;
 using RazorForms.Options;
 
 namespace RazorForms.TagHelpers;
